Move selection function to util
